# Add arm64 assembly implementations for 32 & 64 bits