Here is the arm64 version, tested on RaspberryPi 4 with Linux.

Before

```console
$ go test -bench=. -tags noasm
goos: linux
goarch: arm64
pkg: github.com/tsuna/endian
Benchmark16_1000-4   	  356767	      3354 ns/op
Benchmark32_1000-4   	  305560	      3919 ns/op
Benchmark64_1000-4   	  178471	      6710 ns/op
PASS
ok  	github.com/tsuna/endian	3.600s
```

After

```console
$ go test -bench=.
goos: linux
goarch: arm64
pkg: github.com/tsuna/endian
Benchmark16_1000-4   	  351492	      3359 ns/op
Benchmark32_1000-4   	  356980	      3354 ns/op
Benchmark64_1000-4   	  306332	      3911 ns/op
PASS
ok  	github.com/tsuna/endian	3.585s
```